Removed all routes from index.js and made routes js files.
