+++ views/pages/cardAdd.ejs
@@ -1,3 +1,4 @@
+<link rel='stylesheet' href='/static/css/placeholder.css'>
 <div class="row mt-3">
     <div class="col-sm">
         <!--Spacing--> 
@@ -13,7 +14,7 @@
                 <button class="btn btn-danger m-2" type="submit">Add to Database</button>
             </div>
             <div class="form-group p-2 text-center">
-                <a class="btn btn-secondary m-2" href="/cardAdmin">Go Back to Admin Page</a>   
+                <a class="btn btn-secondary m-2" href="cardLogin/cardAdmin">Go Back to Admin Page</a>   
             </div>
         </form>
     </div>
