+++ views/pages/cardAdmin.ejs
@@ -8,7 +8,7 @@
                 <div class="form-group text-center">
                     <label for="searchword"></label>
                     <small><br> Please enter the exact board number. or</small>
-                    <a class="small ml-1" href="/cardAdminSN">Search by Serial Number</a>
+                    <a class="small ml-1" href="/cardLogin/cardAdminSN">Search by Serial Number</a>
                     <input class="form-control my-3" type="text" name="searchWord" placeholder="Enter Board Number " />
                     <button class="btn btn-danger m-2" type="submit">Search for Board to Edit</button>
                 </div>
@@ -16,7 +16,7 @@
                     <a class="btn btn-success m-2" href="/cardAdd">Add Board to DB</br></a>
                 </div>
                 <div class="form-group text-center">
-                    <a href="/cardLogout" class="btn btn-secondary">Logout</a>
+                    <a href="cardLogout" class="btn btn-secondary">Logout</a>
                 </div>
             </form>
         </div>
