+++ views/pages/cardAdminSN.ejs
@@ -8,7 +8,7 @@
                 <div class="form-group text-center">
                     <label for="searchword"></label>
                     <small><br> Please enter the exact serial number. or</small>
-                    <a class="small ml-1" href="/cardAdmin">Search by Board Number</a>
+                    <a class="small ml-1" href="/cardLogin/cardAdmin">Search by Board Number</a>
                     <input class="form-control my-3" type="text" name="searchWord" placeholder="Enter Serial Number " />
                     <button class="btn btn-danger m-2" type="submit">Search for Board to Edit</button>
                 </div>
