+++ views/pages/cardEdit.ejs
@@ -1,3 +1,4 @@
+<link rel='stylesheet' href='/static/css/placeholder.css'>
 <div class="row mt-3">
     <div class="col-sm">
         <!--Spacing--> 
@@ -15,13 +16,12 @@
                                     <input class="form-control my-3" type="text" name="binNumber" placeholder="Enter bin number" value='<%=item.binNumber%>'/>
                                     <input class="form-control my-3" type="text" name="binLocation" placeholder="Enter Bin Location" value='<%=item.binLocation%>'/>
                                     <button class="btn btn-primary m-2" type="submit">Save Changes</button>
-                                    <a class="btn btn-danger m-2" id="deletebutton" href="/del/<%=item._id%>/delete">Delete Board</a>
                                 </div>
                             </form>
                         </li>
                     <% }); %>
                 </ol>
-                <a class="btn btn-secondary m-2" href="/cardAdmin">Go Back to Admin Page</a>
+                <a class="btn btn-secondary m-2" href="cardLogin/cardAdmin">Go Back to Admin Page</a>
             </div>
         </div>
     <div class="col-sm">
