+++ views/pages/cardSearchSN.ejs
@@ -4,7 +4,7 @@
     </div>
     <div class="col-sm-6">
         
-        <form action='/cardSearchResult' method='POST' name = "form">
+        <form action='/cardSearchResultSN' method='POST' name = "form">
             <div class="form-group p-2">
                 <label class="h5" for='searchword'>Enter a Board Number to Search our Legacy Stock</label> 
                 <small>
