+++ views/pages/cardUpdate.ejs
@@ -38,7 +38,7 @@
                     </li>
                 <% }); %>
             </ol>
-            <a class="btn btn-secondary mt-4" href="/cardAdmin">Go Back to Admin Page</a>
+            <a class="btn btn-secondary mt-4" href="cardLogin/cardAdmin">Go Back to Admin Page</a>
         </div>
     </div>
     <div class="col-sm">
