+++ views/pages/partAdd.ejs
@@ -1,11 +1,10 @@
 <link rel='stylesheet' href='static/css/placeholder.css'>
-
 <div class="row mt-3">
     <div class="col-sm">
         <!--Spacing--> 
     </div>
     <div class="col-sm-6">
-        <form action="/partAdd" method="POST">
+        <form action="partAdd" method="POST">
             <div class="form-group p-2 text-center">
                 <label for="searchword">Enter Information</label>
                 <input class="form-control my-3" type="text" name="stockedAS" placeholder="Part Number" />
@@ -25,7 +24,7 @@
             <div class="form-group p-2 text-center">
                     
             </div>
-                <a class="btn btn-secondary m-2" href="/partAdmin">Go Back to Admin Page</a>
+                <a class="btn btn-secondary m-2" href="/partLogin/partAdmin">Go Back to Admin Page</a>
             </div>
         </form>
     </div>
