+++ views/pages/partEdit.ejs
@@ -1,5 +1,4 @@
-<link rel='stylesheet' href='static/css/placeholder.css'>
-
+<link rel='stylesheet' href='/static/css/placeholder.css'>
 <div class="row mt-3">
     <div class="col-sm">
         <!--Spacing--> 
@@ -9,7 +8,7 @@
             <ol class="list-unstyled">
                 <% response.forEach(function(item){%>
                     <li class="my-3 text-left">
-                        <form action="/partEdit/<%=item._id%>" method="POST">
+                        <form action="partEdit/<%=item._id%>" method="POST">
                             <div class="form-group p-2 text-center">
                             <label for="searchword">Enter Information</label>
                             <input class="form-control my-3" type="text" name="stockedAS" placeholder="Part Number" value='<%=item.stockedAS%>' />
@@ -26,13 +25,13 @@
                             <input class="form-control my-3" type="text" name="cross3" placeholder="Cross 3" value='<%=item.cross3%>'/>
                             <input class="form-control my-3" type="text" name="price" placeholder="List Price" value='<%=item.price%>'/>
                             <button class="btn btn-primary m-2" type="submit">Submit Changes</button>
-                            <a class="btn btn-danger m-2" id="deletebutton" href="/delpart/<%=item._id%>/delete">Delete Part</a>
+                            <a class="btn btn-danger m-2" id="deletebutton" href="/partHome/delpart/<%=item._id%>/delete">Delete Part</a>
                             </div>
                         </form>
                     </li>
                 <% }); %>
             </ol>
-            <a class="btn btn-secondary m-2" href="/partAdmin">Go Back to Admin Page</a>
+            <a class="btn btn-secondary m-2" href="/partLogin/partAdmin">Go Back to Admin Page</a>
         </div>
     </div>
     <div class="col-sm">
