+++ views/pages/partHomeLV.ejs
@@ -3,7 +3,7 @@
         <!--Spacing--> 
     </div>
     <div class="col-sm-6">
-        <form action="/partSearchResultLV" method="POST">
+        <form action="partSearchResultLV" method="POST">
             <div class="form-group p-2">
                 <label class="h5" for="searchword">Enter a Part Number to search our stock </label>
                 <small><br> Don't know your exact part or revision number? <br> You can enter a partial number. EX: 104x123 or</small>
