+++ views/pages/partLogin.ejs
@@ -6,7 +6,7 @@
         <div class="card card-body text-center">
             <h1 class="text-center mb-3">Login</h1>
             	<%- include('../partials/messages'); %>
-            <form action="/partLogin" method="POST">
+            <form action="partLogin/partLogin" method="POST">
 				<div class="form-group">
 					<label for="email">Email</label>
 					<input type="email" id="email" name="email" class="form-control" placeholder="Enter Email"/>
@@ -18,10 +18,10 @@
 				<button type="submit" class="btn btn-primary btn-block">Login</button>
             </form>
             <div class="col-sm p-2">
-              	No Account? <a href="/userRegister">Register</a>
+              	No Account? <a href="partLogin/userRegister">Register</a>
             </div>
             <div class="col-sm">
-              	<a href="/userEdit">Change Password</a>
+              	<a href="partLogin/userEdit">Change Password</a>
             </div>
           </div>
     </div>
@@ -29,6 +29,6 @@
         <!--Spacing--> 
     </div>
 </div>
-<script src="static/scripts/jquery-3.3.1.slim.min.js"></script>
-<script src="static/scripts/popper.min.js"></script>
-<script src="static/scripts/bootstrap.min.js"></script>
+<script src="/static/scripts/jquery-3.3.1.slim.min.js"></script>
+<script src="/static/scripts/popper.min.js"></script>
+<script src="/static/scripts/bootstrap.min.js"></script>
