+++ views/pages/partRequest.ejs
@@ -6,7 +6,7 @@
         <div class="text-center mb-4 h5">Requested Part
             <ol class="list-unstyled">
                 <li class="my-3">
-                    <form action="/quoteRequest" method="POST">
+                    <form action="/partRequest/quoteRequest" method="POST">
                         <label for="searchword">Part Info</label></br>
                         <div class="form-group p-2 text-left">
                             <% if (requestStockedAS) { %>
