+++ views/pages/partRestock.ejs
@@ -6,7 +6,7 @@
         <div class="text-center mb-4 h5">Restock Parts Form
             <ol class="list-unstyled">
                 <li class="my-3">
-                    <form action="/restockOrder" method="POST">
+                    <form action="/partRequest/restockOrder" method="POST">
                         <label for="searchword">Part Info</label></br>
                         <div class="form-group p-2 text-left">
                             <% if (restockAS) { %>
