+++ views/pages/partSearchResult.ejs
@@ -51,13 +51,13 @@
                             <input type='hidden', name='price', value='<%=item.price%>'>
                             <button class= "btn btn-danger mt-2" type='submit'>Request Quote</button>
                         </form>
-                            <form action='/partRestock' method="POST">
+                            <form action='/partRequest/partRestock' method="POST">
                             <input type='hidden', name='stockedAS', value='<%=item.stockedAS%>'>
                             <input type='hidden', name='description1', value='<%=item.description1%>'>
                             <input type='hidden', name='sapNumber', value='<%=item.sapNumber%>'>
                             <button class= "btn btn-secondary mt-2" type='submit'>Restock Order</button>
                         </form>
-                        <!-- <form action='/printLabel' method="POST">
+                        <!-- <form action='/partHome/printLabel' method="POST">
                             <input type='hidden', name='stockedAS', value='<%=item.stockedAS%>'>
                             <input type='hidden', name='description1', value='<%=item.description1%>'>
                             <input type='hidden', name='sapNumber', value='<%=item.sapNumber%>'>
@@ -68,7 +68,7 @@
                     </li>
                 <% }); %>
             </ol>
-            <a class="btn btn-primary mt-4 text-white" href="/requestNewPart">Request Part not on list</a> 
+            <a class="btn btn-primary mt-4 text-white" href="/partRequest/requestNewPart">Request Part not on list</a> 
         <div>
             <a class="btn btn-secondary mt-4 text-white" href="/partHome">Go Back to Parts Search</a>
         </div>
