+++ views/pages/requestNewPart.ejs
@@ -6,7 +6,7 @@
         <div class="text-center mb-4 h5">Requested New Part
             <ol class="list-unstyled">
                 <li class="my-3">
-                    <form action="/requestNewPart" method="POST">
+                    <form action="/partRequest/requestNewPart" method="POST">
                         <div class="form-group p-2 text-center">
                             <label for="searchword">New Part Info</label>
                             <input class="form-control my-3" type="text" name="stockedAS" placeholder="Part Number" />
