+++ views/pages/requestSent.ejs
@@ -6,7 +6,7 @@
         <p>Your request has been submitted. Our team is working on your quote. If you need to make any changes to your request
             Please email <a href="mailto: US-CustomerService.Louisville@abb.com?subject=Modification of Quote Request">Customer Service</a> 
         </p>
-        <a class="btn btn-secondary btn-block mt-4" href="partHome">Back to Parts Search</a>
+        <a class="btn btn-secondary btn-block mt-4" href="/partHome">Back to Parts Search</a>
     </div>
     <div class="col-sm">
         <!--Spacing-->
