+++ views/pages/restockSearchResults.ejs
@@ -25,12 +25,12 @@
                                 <strong class="pr-2">Requested By: </strong> <%= item.requestor %><br>
                             <% } %>
 
-                            <a class="btn btn-danger btn-sm mt-2 text-white" id="deletebutton" href="/restockDelete/<%=item._id%>/delete">Remove From List</a> 
+                            <a class="btn btn-danger btn-sm mt-2 text-white" id="deletebutton" href="/partRequest/restockDelete/<%=item._id%>/delete">Remove From List</a> 
                             <hr>
                         </form>
                     </li>
                 <% }); %>
-                <a class="btn btn-secondary btn-block mt-4 text-white" href="/partAdmin">Go Back to Admin Page</a>
+                <a class="btn btn-secondary btn-block mt-4 text-white" href="/partLogin/partAdmin">Go Back to Admin Page</a>
             </ol>
         </div>
     </div>
