+++ views/pages/userRegister.ejs
@@ -6,7 +6,7 @@
 		<div class="card card-body text-center">
 			<h1 class="text-center mb-3">Register</h1>
 			<%- include ("../partials/messages"); %>
-			<form action="/userRegister" method="POST">
+			<form action="userRegister" method="POST">
 				<div class="form-group">
 					<label for="nameFirst">First Name</label>
 					<input
@@ -67,13 +67,13 @@
 				</button>
 				<a class="btn btn-secondary m-2" href="/">Back to home page</a>
 			</form>
-			<p class="lead">Have An Account?<br><a href="/loginPart">Parts Login</a><br><a href="/loginCard">Legacy Login</a></p>
+			<p class="lead">Have An Account?<br><a href="/partLogin">Parts Login</a><br><a href="/cardLogin">Legacy Login</a></p>
 		</div>
     </div>
     <div class="col-sm">
         <!--Spacing--> 
     </div>
 </div>
-<script src="static/scripts/jquery-3.3.1.slim.min.js"></script>
-<script src="static/scripts/popper.min.js"></script>
-<script src="static/scripts/bootstrap.min.js"></script>
+<script src="/static/scripts/jquery-3.3.1.slim.min.js"></script>
+<script src="/static/scripts/popper.min.js"></script>
+<script src="/static/scripts/bootstrap.min.js"></script>
