+++ views/partials/head.ejs
@@ -1,8 +1,8 @@
 <meta charset=utf-8>
 <meta name="viewport" content="width=device-width,initial-scale=1.0">
-<link rel='stylesheet' href='static/css/bootstrap.min.css'>
-<link rel='stylesheet' href='static/css/print.css'>
-<link rel='icon' type='image/png' href='static/photos/abb-logo-square.png'>
+<link rel='stylesheet' href='/static/css/bootstrap.min.css'>
+<link rel='stylesheet' href='/static/css/print.css'>
+<link rel='icon' type='image/png' href='/static/photos/abb-logo-square.png'>
 <meta type="text/css" media="print">
 <title>
     ABB Louisville - MOSE
