+++ views/partials/header.ejs
@@ -1,6 +1,6 @@
 <div class="row mt-2">
     <div class="col-sm-5 h5">
-        <img class="img img-fluid mr-3" src="static/photos/abb.png">
+        <img class="img img-fluid mr-3" src="/static/photos/abb.png">
         Louisville
     </div>
     <div class="col-sm text-right">
